made the gameplay loop
Developed the pipes. Added a collision detection to the ground and the pipes to enable lose condition. Stop game if collision is detected. added a start game condition and a reset condition upon loss.

Attempted to add a counter to count the pipes that have been past. needs more looking at